[extension] Deprecate extension/auth in favor of extension/extensionauth
