+++ cmd/builder/internal/builder/main_test.go
@@ -76,8 +76,8 @@ var replaceModules = []string{
 	"/exporter/otlpexporter",
 	"/exporter/otlphttpexporter",
 	"/extension",
-	"/extension/auth",
-	"/extension/auth/authtest",
+	"/extension/extensionauth",
+	"/extension/extensionauth/extensionauthtest",
 	"/extension/extensioncapabilities",
 	"/extension/extensiontest",
 	"/extension/zpagesextension",
