+++ cmd/otelcorecol/builder-config.yaml
@@ -76,8 +76,8 @@ replaces:
   - go.opentelemetry.io/collector/exporter/otlpexporter => ../../exporter/otlpexporter
   - go.opentelemetry.io/collector/exporter/otlphttpexporter => ../../exporter/otlphttpexporter
   - go.opentelemetry.io/collector/extension => ../../extension
-  - go.opentelemetry.io/collector/extension/auth => ../../extension/auth
-  - go.opentelemetry.io/collector/extension/auth/authtest => ../../extension/auth/authtest
+  - go.opentelemetry.io/collector/extension/extensionauth => ../../extension/extensionauth
+  - go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest => ../../extension/extensionauth/extensionauthtest
   - go.opentelemetry.io/collector/extension/extensioncapabilities => ../../extension/extensioncapabilities
   - go.opentelemetry.io/collector/extension/extensiontest => ../../extension/extensiontest
   - go.opentelemetry.io/collector/extension/memorylimiterextension => ../../extension/memorylimiterextension
