+++ config/confighttp/confighttp_test.go
@@ -30,8 +30,8 @@ import (
 	"go.opentelemetry.io/collector/config/configcompression"
 	"go.opentelemetry.io/collector/config/configopaque"
 	"go.opentelemetry.io/collector/config/configtls"
-	"go.opentelemetry.io/collector/extension/auth"
-	"go.opentelemetry.io/collector/extension/auth/authtest"
+	"go.opentelemetry.io/collector/extension/extensionauth"
+	"go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest"
 )
 
 type customRoundTripper struct{}
@@ -54,7 +54,7 @@ var (
 func TestAllHTTPClientSettings(t *testing.T) {
 	host := &mockHost{
 		ext: map[component.ID]component.Component{
-			testAuthID: &authtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
+			testAuthID: &extensionauthtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
 		},
 	}
 
@@ -180,7 +180,7 @@ func TestAllHTTPClientSettings(t *testing.T) {
 func TestPartialHTTPClientSettings(t *testing.T) {
 	host := &mockHost{
 		ext: map[component.ID]component.Component{
-			testAuthID: &authtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
+			testAuthID: &extensionauthtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
 		},
 	}
 
@@ -348,7 +348,7 @@ func TestHTTPClientSettingWithAuthConfig(t *testing.T) {
 			shouldErr: false,
 			host: &mockHost{
 				ext: map[component.ID]component.Component{
-					mockID: &authtest.MockClient{
+					mockID: &extensionauthtest.MockClient{
 						ResultRoundTripper: &customRoundTripper{},
 					},
 				},
@@ -363,7 +363,7 @@ func TestHTTPClientSettingWithAuthConfig(t *testing.T) {
 			shouldErr: true,
 			host: &mockHost{
 				ext: map[component.ID]component.Component{
-					mockID: &authtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
+					mockID: &extensionauthtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
 				},
 			},
 		},
@@ -385,7 +385,7 @@ func TestHTTPClientSettingWithAuthConfig(t *testing.T) {
 			shouldErr: false,
 			host: &mockHost{
 				ext: map[component.ID]component.Component{
-					mockID: &authtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
+					mockID: &extensionauthtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
 				},
 			},
 		},
@@ -399,7 +399,7 @@ func TestHTTPClientSettingWithAuthConfig(t *testing.T) {
 			shouldErr: false,
 			host: &mockHost{
 				ext: map[component.ID]component.Component{
-					mockID: &authtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
+					mockID: &extensionauthtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
 				},
 			},
 		},
@@ -413,7 +413,7 @@ func TestHTTPClientSettingWithAuthConfig(t *testing.T) {
 			shouldErr: false,
 			host: &mockHost{
 				ext: map[component.ID]component.Component{
-					mockID: &authtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
+					mockID: &extensionauthtest.MockClient{ResultRoundTripper: &customRoundTripper{}},
 				},
 			},
 		},
@@ -426,7 +426,7 @@ func TestHTTPClientSettingWithAuthConfig(t *testing.T) {
 			shouldErr: true,
 			host: &mockHost{
 				ext: map[component.ID]component.Component{
-					mockID: &authtest.MockClient{
+					mockID: &extensionauthtest.MockClient{
 						ResultRoundTripper: &customRoundTripper{}, MustError: true,
 					},
 				},
@@ -825,8 +825,8 @@ func TestHttpCorsWithSettings(t *testing.T) {
 
 	host := &mockHost{
 		ext: map[component.ID]component.Component{
-			mockID: auth.NewServer(
-				auth.WithServerAuthenticate(func(ctx context.Context, _ map[string][]string) (context.Context, error) {
+			mockID: extensionauth.NewServer(
+				extensionauth.WithServerAuthenticate(func(ctx context.Context, _ map[string][]string) (context.Context, error) {
 					return ctx, errors.New("Settings failed")
 				}),
 			),
@@ -1137,8 +1137,8 @@ func TestServerAuth(t *testing.T) {
 
 	host := &mockHost{
 		ext: map[component.ID]component.Component{
-			mockID: auth.NewServer(
-				auth.WithServerAuthenticate(func(ctx context.Context, _ map[string][]string) (context.Context, error) {
+			mockID: extensionauth.NewServer(
+				extensionauth.WithServerAuthenticate(func(ctx context.Context, _ map[string][]string) (context.Context, error) {
 					authCalled = true
 					return ctx, nil
 				}),
@@ -1188,8 +1188,8 @@ func TestFailedServerAuth(t *testing.T) {
 	}
 	host := &mockHost{
 		ext: map[component.ID]component.Component{
-			mockID: auth.NewServer(
-				auth.WithServerAuthenticate(func(ctx context.Context, _ map[string][]string) (context.Context, error) {
+			mockID: extensionauth.NewServer(
+				extensionauth.WithServerAuthenticate(func(ctx context.Context, _ map[string][]string) (context.Context, error) {
 					return ctx, errors.New("Settings failed")
 				}),
 			),
@@ -1369,8 +1369,8 @@ func TestAuthWithQueryParams(t *testing.T) {
 
 	host := &mockHost{
 		ext: map[component.ID]component.Component{
-			mockID: auth.NewServer(
-				auth.WithServerAuthenticate(func(ctx context.Context, sources map[string][]string) (context.Context, error) {
+			mockID: extensionauth.NewServer(
+				extensionauth.WithServerAuthenticate(func(ctx context.Context, sources map[string][]string) (context.Context, error) {
 					require.Len(t, sources, 1)
 					assert.Equal(t, "1", sources["auth"][0])
 					authCalled = true
