+++ config/confighttp/xconfighttp/go.mod
@@ -32,7 +32,7 @@ require (
 	go.opentelemetry.io/collector/config/configopaque v1.26.0 // indirect
 	go.opentelemetry.io/collector/config/configtls v1.26.0 // indirect
 	go.opentelemetry.io/collector/extension v0.120.0 // indirect
-	go.opentelemetry.io/collector/extension/auth v0.120.0 // indirect
+	go.opentelemetry.io/collector/extension/extensionauth v0.120.0 // indirect
 	go.opentelemetry.io/collector/pdata v1.26.0 // indirect
 	go.opentelemetry.io/otel v1.34.0 // indirect
 	go.opentelemetry.io/otel/metric v1.34.0 // indirect
@@ -54,7 +54,7 @@ replace go.opentelemetry.io/collector/client => ../../../client
 
 replace go.opentelemetry.io/collector/consumer => ../../../consumer
 
-replace go.opentelemetry.io/collector/extension/auth/authtest => ../../../extension/auth/authtest
+replace go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest => ../../../extension/extensionauth/extensionauthtest
 
 replace go.opentelemetry.io/collector/component/componenttest => ../../../component/componenttest
 
@@ -62,7 +62,7 @@ replace go.opentelemetry.io/collector/config/configauth => ../../configauth
 
 replace go.opentelemetry.io/collector/pdata => ../../../pdata
 
-replace go.opentelemetry.io/collector/extension/auth => ../../../extension/auth
+replace go.opentelemetry.io/collector/extension/extensionauth => ../../../extension/extensionauth
 
 replace go.opentelemetry.io/collector/config/configopaque => ../../configopaque
