+++ extension/extensionauth/doc.go
@@ -4,4 +4,4 @@
 // Package auth implements the configuration settings to
 // ensure authentication on incoming requests, and allows
 // exporters to add authentication on outgoing requests.
-package auth // import "go.opentelemetry.io/collector/extension/auth"
+package extensionauth // import "go.opentelemetry.io/collector/extension/extensionauth"
