+++ extension/extensionauth/extensionauthtest/go.mod
@@ -1,15 +1,20 @@
-// Deprecated: use go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest instead.
-module go.opentelemetry.io/collector/extension/auth/authtest
+module go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest
 
 go 1.23.0
 
-require go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest v0.0.0-00010101000000-000000000000
+require (
+	github.com/stretchr/testify v1.10.0
+	go.opentelemetry.io/collector/component v0.120.0
+	go.opentelemetry.io/collector/extension/extensionauth v0.120.0
+	go.uber.org/goleak v1.3.0
+	google.golang.org/grpc v1.70.0
+)
 
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
-	go.opentelemetry.io/collector/component v0.120.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	go.opentelemetry.io/collector/extension v0.120.0 // indirect
-	go.opentelemetry.io/collector/extension/extensionauth v0.120.0 // indirect
 	go.opentelemetry.io/collector/pdata v1.26.0 // indirect
 	go.opentelemetry.io/otel v1.34.0 // indirect
 	go.opentelemetry.io/otel/metric v1.34.0 // indirect
@@ -20,13 +25,11 @@ require (
 	golang.org/x/sys v0.29.0 // indirect
 	golang.org/x/text v0.21.0 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20241202173237-19429a94021a // indirect
-	google.golang.org/grpc v1.70.0 // indirect
 	google.golang.org/protobuf v1.36.5 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
 
-replace go.opentelemetry.io/collector/extension/extensionauth => ../../extensionauth
-
-replace go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest => ../../extensionauth/extensionauthtest
+replace go.opentelemetry.io/collector/extension/extensionauth => ..
 
 replace go.opentelemetry.io/collector/component => ../../../component
