+++ otelcol/otelcoltest/go.mod
@@ -201,7 +201,7 @@ replace go.opentelemetry.io/collector/config/confighttp => ../../config/confight
 
 replace go.opentelemetry.io/collector/client => ../../client
 
-replace go.opentelemetry.io/collector/extension/auth => ../../extension/auth
+replace go.opentelemetry.io/collector/extension/extensionauth => ../../extension/extensionauth
 
 replace go.opentelemetry.io/collector/otelcol => ../
 
@@ -215,7 +215,7 @@ replace go.opentelemetry.io/collector/internal/fanoutconsumer => ../../internal/
 
 replace go.opentelemetry.io/collector/extension/extensiontest => ../../extension/extensiontest
 
-replace go.opentelemetry.io/collector/extension/auth/authtest => ../../extension/auth/authtest
+replace go.opentelemetry.io/collector/extension/extensionauth/extensionauthtest => ../../extension/extensionauth/extensionauthtest
 
 replace go.opentelemetry.io/collector/extension/xextension => ../../extension/xextension
